fixed delete button on add gear

+++ public/js/delBtn.js
@@ -1,16 +1,21 @@
 const delBtns = document.querySelectorAll(".del-btn")
 
-delBtns.forEach(btn=>{
-    btn.addEventListener("click",e=>{
+delBtns.forEach(btn => {
+    btn.addEventListener("click", e => {
         const delId = e.target.getAttribute("data-id");
-        fetch(`/api/gears/${delId}`,{
-            method:"DELETE"
-        }).then(res=>{
-            if(res.ok){
-               location.reload()
-            } else {
-                alert("trumpet sound")
-            }
-        })
+        
+        if (delId) {
+            console.log(delId)
+            fetch(`/api/gears/${delId}`, {
+                method: "DELETE"
+            }).then(res => {
+                if (res.ok) {
+                    location.reload()
+                } else {
+                    alert("trumpet sound")
+                }
+            })
+        }
+
     })
 })
